requirements.txt
aiokafka
pydantic
pydantic_settings
aiosmtplib
fastapi[all]
python-dotenv